Edit find Command
