+++ src/test/java/seedu/address/logic/commands/AddTaskCommandTest.java
@@ -3,7 +3,6 @@ package seedu.address.logic.commands;
 import static seedu.address.testutil.Assert.assertThrows;
 
 import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
@@ -119,7 +118,7 @@ public class AddTaskCommandTest {
         }
 
         @Override
-        public ArrayList<Task> findTask(String input) {
+        public String findTask(String input) {
             throw new AssertionError("This method should not be called."); }
     }
