+++ src/test/java/seedu/address/logic/commands/DeleteTaskCommandTest.java
@@ -3,7 +3,6 @@ package seedu.address.logic.commands;
 import static seedu.address.testutil.Assert.assertThrows;
 
 import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
@@ -113,7 +112,7 @@ public class DeleteTaskCommandTest {
         }
 
         @Override
-        public ArrayList<Task> findTask(String input) {
+        public String findTask(String input) {
             throw new AssertionError("This method should not be called."); }
     }
 }
